+++ CHANGELOG.md
@@ -24,7 +24,10 @@ Adding a requirement of a major version of a dependency is breaking a contract.
 Dropping a requirement of a major version of a dependency is a new contract.
 
 ## [Unreleased]
-[Unreleased]: https://bitbucket.org/atlassian/infrastructure/branches/compare/master%0Drelease-4.1.0
+[Unreleased]: https://bitbucket.org/atlassian/infrastructure/branches/compare/master%0Drelease-4.2.0
+
+## [4.2.0] - 2018-12-04
+[4.2.0]: https://bitbucket.org/atlassian/infrastructure/branches/compare/release-4.2.0%0Drelease-4.1.0
 
 ### Added
 - Support providing collectd configurations via `collectdConfigs` in `JiraNodeConfig`. Unblock [JPERF-285].